requirements.txt
# ─── Core ─────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ─── Data Processing ──────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.26.0
beautifulsoup4>=4.12.0

# ─── Embeddings ───────────────────────────────────────────────────────────────
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.2.0

# ─── Vector Store ─────────────────────────────────────────────────────────────
chromadb>=0.5.0
# Alternative: faiss-cpu>=1.8.0

# ─── LLM / Generation ─────────────────────────────────────────────────────────
openai>=1.30.0           # optional: for API-based LLM
ollama>=0.2.0            # optional: for local models

# ─── RAG Framework (optional) ─────────────────────────────────────────────────
# langchain>=0.2.0
# llama-index>=0.10.0

# ─── Experiment Tracking ──────────────────────────────────────────────────────
mlflow>=2.13.0

# ─── Evaluation ───────────────────────────────────────────────────────────────
rank-bm25>=0.2.2

# ─── Demo Interface ───────────────────────────────────────────────────────────
streamlit>=1.35.0

# ─── API Backend (React frontend) ─────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# ─── Notebooks ────────────────────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.29.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=5.0.0
